Further startup optimizations.
Set gc-cons-threshold in early-init.el.  But keep it also set in
init.el, for compatibility reasons.  Settings it back to the default
value should be done *after* the message is printed, otherwise the GC
might fire.

In addition, set file-name-handler-alist to nil, to further decrease
the startup time.  This might have a negative impact.  Need to monitor this.